SpatialRings-class {rgeos}	R Documentation

Class "SpatialRings"

Description

class to hold linear ring topology (without attributes)

Objects from the Class

Objects can be created by calls to the function SpatialRings

Slots

rings:

Object of class "list"; list elements are all of class Ring-class

bbox:

Object of class "matrix"; see Spatial-class

proj4string:

Object of class "CRS"; see CRS-class

Extends

Class "Spatial", directly.

Methods

Methods defined with class "SpatialRings" in the signature:

[

signature(obj = "SpatialRings"): select subset of (sets of) rings; NAs are not permitted in the row index

plot

signature(x = "SpatialRings", y = "missing"): plot rings in SpatialRings object

bbox

signature(obj = "SpatialRings"): retrieves the bbox element

coordinates

signature(object = "SpatialRings"): retrieves the coords element from Ring objects in rings slot

coordnames

signature(object = "SpatialRings"): retrieves coordinate names

row.names

signature(object = "SpatialRings"): retrieves the ID element from Ring objects in rings slot

spChFIDs

signature(obj="SpatialRings", x="character"): replaces ID element

coerce

signature(from = "SpatialRings", to = "SpatialPoints"): ...

Author(s)

Colin Rundel

See Also

SpatialRings Ring-class

Examples

 
	#NONE 

[Package rgeos version 0.5-2 Index]