SpatialRingsDataFrame-class {rgeos}	R Documentation

Class "SpatialRingsDataFrame"

Description

class to hold linear ring topology (without attributes)

Objects from the Class

Objects can be created by calls to the function SpatialRingsDataFrame

Slots

data:

Object of class "data.frame"; attribute table

rings:

Object of class "list"; list elements are all of class Ring-class

bbox:

Object of class "matrix"; see Spatial-class

proj4string:

Object of class "CRS"; see CRS-class

Extends

Class "SpatialRings", directly. Class "Spatial", by class "SpatialRings".

Methods

Methods defined with class "SpatialRingsDataFrame" in the signature:

[

signature(obj = "SpatialRingsDataFrame"): select subset of (sets of) rings; NAs are not permitted in the row index

plot

signature(x = "SpatialRingsDataFrame", y = "missing"): plot rings in SpatialRingsDataFrame object

bbox

signature(obj = "SpatialRingsDataFrame"): retrieves the bbox element

coordinates

signature(object = "SpatialRingsDataFrame"): retrieves the coords element from Ring objects in rings slot

coordnames

signature(object = "SpatialRingsDataFrame"): retrieves coordinate names

row.names

signature(object = "SpatialRingsDataFrame"): retrieves the ID element from Ring objects in rings slot

spChFIDs

signature(obj="SpatialRingsDataFrame", x="character"): replaces ID element

names

signature(object = "SpatialRingsDataFrame"): retrieves names from data element

dim

signature(object = "SpatialRingsDataFrame"): retrieves dimensions of data element

coerce

signature(from = "SpatialRingsDataFrame", to = "SpatialPoints"): ...

coerce

signature(from = "SpatialRingsDataFrame", to = "SpatialRings"): ...

coerce

signature(from = "SpatialRingsDataFrame", to = "data.frame"): ...

Author(s)

Colin Rundel

See Also

SpatialRingsDataFrame Ring-class SpatialRings-class

Examples

 
	#NONE 

[Package rgeos version 0.4-1 Index]