gpc.poly.nohole-class {rgeos}	R Documentation

Class "gpc.poly.nohole"

Description

A class for representing polygons with multiple contours but without holes.

Objects from the Class

Objects can be created by calls of the form ‘new("gpc.poly.nohole", ...) or by calling read.polyfile’.

Slots

pts

Object of class “list”. See the help for “gpc.poly” for details.

Extends

Class “gpc.poly”, directly.

Methods

coerce

signature(from = "numeric", to = "gpc.poly.nohole"): ...

Note

This class is identical to “"gpc.poly"” and is needed because the file formats for polygons without holes is slightly different from the file format for polygons with holes. For a “gpc.poly.nohole” object, the hole flag for each contour is always FALSE.

Also, write.polyfile will write the correct file format, depending on whether the object is of class “gpc.poly” or “gpc.poly.nohole”.

Author(s)

Roger D. Peng

See Also

gpc.poly-class

Examples

## None

[Package rgeos version 0.5-2 Index]