SpatialCollections {rgeos}	R Documentation

create SpatialCollections

Description

create object of class SpatialCollections

Usage

 
SpatialCollections(points=NULL, lines=NULL, rings=NULL, polygons=NULL,
 plotOrder=c(4,3,2,1), proj4string=CRS(as.character(NA)))

Arguments

points
list with objects of class SpatialPoints-class

lines
list with objects of class SpatialLines-class

rings
list with objects of class SpatialRings-class

polygons
list with objects of class SpatialPolygons-class

plotOrder
numeric vector of length 4 that determines the order in which the geometries will be plotted. By default polygons will be plotted followed by rings, then lines and finally points.

proj4string
Object of class "CRS" holding a valid proj4 string

Value

SpatialCollections returns object of class SpatialCollections

See Also

SpatialCollections-class SpatialPoints-class SpatialLines-class SpatialRings-class SpatialPolygons-class

[Package rgeos version 0.4-1 Index]