gInterpolate {rgeos}	R Documentation

Interpolate Points along Line Geometry

Description

Return points at specified distances along a line.

Usage

gInterpolate(spgeom, d, normalized = FALSE)

Arguments

spgeom
SpatialLines or SpatialLinesDataFrame object

d
Numeric vector specifying the distance along the line geometry

normalized
Logical determining if normalized distances should be used

Details

If normalized=TRUE, the distances will be interpreted as fractions of the line length.

Value

SpatialPoints object

Author(s)

Rainer Stuetz

See Also

gInterpolate

Examples

gInterpolate(readWKT("LINESTRING(25 50, 100 125, 150 190)"),
             d=seq(0, 1, by = 0.2), normalized = TRUE)

[Package rgeos version 0.5-2 Index]