gProject {rgeos}	R Documentation

Project Points to Line Geometry

Description

Return distances along geometry to points nearest the specified points.

Usage

gProject(spgeom, sppoint, normalized = FALSE)

Arguments

spgeom
SpatialLines or SpatialLinesDataFrame object

sppoint
SpatialPoints or SpatialPointsDataFrame object

normalized
Logical determining if normalized distances should be used

Details

If normalized=TRUE, distances normalized to the length of the geometry are returned, i.e., values between 0 and 1.

Value

a numeric vector containing the distances along the line to points nearest to the specified points

Author(s)

Rainer Stuetz

See Also

gInterpolate

Examples

l <- readWKT("LINESTRING(0 1, 3 4, 5 6)")
p1 <- readWKT("MULTIPOINT(3 2, 3 5)")
frac <- gProject(l, p1)
p2 <- gInterpolate(l, frac)
plot(l, axes=TRUE)
plot(p1, col = "blue", add = TRUE)
plot(p2, col = "red", add = TRUE)

[Package rgeos version 0.5-2 Index]