gIsEmpty {rgeos}	R Documentation

Is Geometry Empty?

Description

Tests if the given geometry is empty

Usage

gIsEmpty(spgeom, byid = FALSE) 

Arguments

spgeom
sp object as defined in package sp

byid
Logical determining if the function should be applied across subgeometries (TRUE) or the entire object (FALSE)

Details

Because no sp Spatial object may be empty, the function exists but cannot work, as readWKT is not permitted to create an empty object.

Value

Returns TRUE if the given geometry is empty, FALSE otherwise.

Author(s)

Roger Bivand & Colin Rundel

See Also

gIsRing gIsSimple gIsValid

Examples

try(gIsEmpty(readWKT("POINT EMPTY")))
gIsEmpty(readWKT("POINT(1 1)"))
try(gIsEmpty(readWKT("LINESTRING EMPTY")))
gIsEmpty(readWKT("LINESTRING(0 0,1 1)"))
try(gIsEmpty(readWKT("POLYGON EMPTY")))
gIsEmpty(readWKT("POLYGON((0 0,1 0,1 1,0 1,0 0))"))

[Package rgeos version 0.4-1 Index]