Geometry Intersections

Function for determining the intersection between the two given geometries

gIntersection(spgeom1, spgeom2, byid=FALSE, id=NULL, drop_not_poly,
 drop_lower_td=FALSE, unaryUnion_if_byid_false=TRUE, checkValidity=FALSE)

Arguments

spgeom1, spgeom2
sp objects as defined in package sp

byid
Logical vector determining if the function should be applied across ids (TRUE) or the entire object (FALSE) for spgeom1 and spgeom2

id
Character vector defining id labels for the resulting geometries, if unspecified returned geometries will be labeled based on their parent geometries' labels.

drop_not_poly
deprecated argument, use drop_lower_td

drop_lower_td
default FALSE; if TRUE, objects will be dropped from output GEOMETRYCOLLECTION objects to simplify output if their topological dimension is less than the minimum topological dinension of the input objects.

unaryUnion_if_byid_false
default TRUE; if byid takes a FALSE in either position, the subgeometries are combined first to avoid possible topology exceptions (change in 0.3-13, previous behaviour did not combine subgeometries, and may be achieved by setting this argument FALSE

checkValidity
default FALSE; error meesages from GEOS do not say clearly which object fails if a topology exception is encountered. If this argument is TRUE, gIsValid is run on each in turn in an environment in which object names are available. If objects are invalid, this is reported and those affected are named

Details

Returns all spatial intersections as sp objects of the appropriate class. If the geometries do not intersect then an empty geometry is returned.

Note

Error messages from GEOS, in particular topology exceptions, report 0-based object order, so geom 0 is spgeom1, and geom 1 is spgeom2.

See also

Examples

if (require(maptools)) {
xx <- readShapeSpatial(system.file("shapes/fylk-val-ll.shp", package="maptools")[1],
 proj4string=CRS("+proj=longlat +datum=WGS84"))
bbxx <- bbox(xx)
wdb_lines <- system.file("share/wdb_borders_c.b", package="maptools")
xxx <- Rgshhs(wdb_lines, xlim=bbxx[1,], ylim=bbxx[2,])$SP
res <-gIntersection(xx, xxx)
plot(xx, axes=TRUE)
plot(xxx, lty=2, add=TRUE)
plot(res, add=TRUE, pch=16,col='red')
}

#> Warning: readShapeSpatial is deprecated; use rgdal::readOGR or sf::st_read

#> Warning: readShapeLines is deprecated; use rgdal::readOGR or sf::st_read

#> Data are line data

#> Error in as.double(y): cannot coerce type 'S4' to vector of type 'double'

pol <- readWKT(paste("POLYGON((-180 -20, -140 55, 10 0, -140 -60, -180 -20),",
 "(-150 -20, -100 -10, -110 20, -150 -20))"))
library(sp)
GT <- GridTopology(c(-175, -85), c(10, 10), c(36, 18))
gr <- as(as(SpatialGrid(GT), "SpatialPixels"), "SpatialPolygons")
try(res <- gIntersection(pol, gr, byid=TRUE))

#> output subgeometry 88, row.name: 1 g365
#> subsubgeometry 0: Point
#> subsubgeometry 1: Polygon
#> Error in RGEOSBinTopoFunc(spgeom1, spgeom2, byid, id, drop_lower_td, unaryUnion_if_byid_false,  : 
#>   Geometry collections may not contain other geometry collections

res <- gIntersection(pol, gr, byid=TRUE, drop_lower_td=TRUE)
# Robert Hijmans difficult intersection case
load(system.file("test_cases/polys.RData", package="rgeos"))
try(res <- gIntersection(a, b, byid=TRUE))

#> output subgeometry 1, row.name: 109 294
#> subsubgeometry 0: LineString
#> subsubgeometry 1: Polygon
#> subsubgeometry 2: Polygon
#> Error in RGEOSBinTopoFunc(spgeom1, spgeom2, byid, id, drop_lower_td, unaryUnion_if_byid_false,  : 
#>   Geometry collections may not contain other geometry collections

res <- gIntersection(a, b, byid=TRUE, drop_lower_td=TRUE)
unlist(sapply(slot(res, "polygons"), function(p) sapply(slot(p, "Polygons"), slot, "area")))

#>  [1] 1.700212e+06 3.135483e-08 9.982614e-07 2.087862e-06 1.434470e-06
#>  [6] 3.926317e-07 5.350546e-07 7.269743e+06 3.117775e+05 1.212778e+03
#> [11] 1.424050e+06

# example from Pius Korner 2015-10-25
poly1 <- SpatialPolygons(list(Polygons(list(Polygon(coords=matrix(c(0, 0, 2, 2, 0, 1, 1, 0),
 ncol=2, byrow=FALSE))), ID=c("a")), Polygons(list(Polygon(coords=matrix(c(0, 0, 2, 2, 2, 3, 3, 2),
 ncol=2, byrow=FALSE))), ID=c("b"))))
poly2 <- SpatialPolygons(list(Polygons(list(Polygon(coords=matrix(c(0, 0, 2, 2,
 1, 1, 1, 3, 3, 0, 0, 2), ncol=2, byrow=FALSE))), ID=c("c"))))
plot(poly1, border="orange")

#> Error in as.double(y): cannot coerce type 'S4' to vector of type 'double'

plot(poly2, border="blue", add=TRUE, lty=2, density=8, angle=30, col="blue")

#> Error in as.double(y): cannot coerce type 'S4' to vector of type 'double'

gI <- gIntersection(poly1, poly2, byid=TRUE, drop_lower_td=TRUE)
plot(gI, add=TRUE, border="red", lwd=3)

#> Error in as.double(y): cannot coerce type 'S4' to vector of type 'double'

oT <- get_RGEOS_polyThreshold()
oW <- get_RGEOS_warnSlivers()
oD <- get_RGEOS_dropSlivers()
set_RGEOS_polyThreshold(1e-3)
set_RGEOS_warnSlivers(TRUE)
res1 <- gIntersection(a, b, byid=TRUE, drop_lower_td=TRUE)

#> Warning: 1: MultiPolygon object 109 294 area 1.02962e-06

#> Warning: 2: MultiPolygon object 202 294 area 3.91496e-06

#> Warning: 3: Polygon object 208 294 area 5.35055e-07

#> Warning: Exterior ring 0 of object 109 294 area 3.13548e-08

#> Warning: Exterior ring 1 of object 109 294 area 9.98261e-07

#> Warning: Exterior ring 0 of object 202 294 area 2.08786e-06

#> Warning: Exterior ring 1 of object 202 294 area 1.43447e-06

#> Warning: Exterior ring 2 of object 202 294 area 3.92632e-07

#> Warning: Exterior ring 0 of object 208 294 area 5.35055e-07

unlist(sapply(slot(res1, "polygons"), function(p) sapply(slot(p, "Polygons"), slot, "area")))

#>  [1] 1.700212e+06 3.135483e-08 9.982614e-07 2.087862e-06 1.434470e-06
#>  [6] 3.926317e-07 5.350546e-07 7.269743e+06 3.117775e+05 1.212778e+03
#> [11] 1.424050e+06

set_RGEOS_dropSlivers(TRUE)
res2 <- gIntersection(a, b, byid=TRUE, drop_lower_td=TRUE)

#> Warning: 1: MultiPolygon object 109 294 area 1.02962e-06

#> Warning: 2: MultiPolygon object 202 294 area 3.91496e-06

#> Warning: 3: Polygon object 208 294 area 5.35055e-07

unlist(sapply(slot(res2, "polygons"), function(p) sapply(slot(p, "Polygons"), slot, "area")))

#> [1] 1700211.935 7269743.131  311777.515    1212.778 1424050.306

set_RGEOS_dropSlivers(FALSE)
oo <- gUnaryUnion(res1, c(rep("1", 3), "2", "3", "4"))

#> Warning: Exterior ring 1 of object 1 area 2.99219e-06

#> Warning: 1: Polygon object 2 area 0

#> Warning: Exterior ring 0 of object 2 area 0

unlist(sapply(slot(oo, "polygons"), function(p) sapply(slot(p, "Polygons"), slot, "area")))

#> [1] 1.700212e+06 2.992188e-06 0.000000e+00 7.269743e+06 3.117775e+05
#> [6] 1.212778e+03

ooo <- gIntersection(b, oo, byid=TRUE)

#> Warning: Exterior ring 1 of object 294 1 area 2.99219e-06

gArea(ooo, byid=TRUE)

#>     294 1     294 3     294 4 
#> 1700211.9 7269743.1  312990.3 

unlist(sapply(slot(ooo, "polygons"), function(p) sapply(slot(p, "Polygons"), slot, "area")))

#> [1] 1.700212e+06 2.992188e-06 7.269743e+06 3.117775e+05 1.212778e+03

set_RGEOS_dropSlivers(TRUE)
ooo <- gIntersection(b, oo, byid=TRUE)

#> Warning: Exterior ring 1 of object 294 1 area 2.99219e-06

gArea(ooo, byid=TRUE)

#>     294 1     294 3     294 4 
#> 1700211.9 7269743.1  312990.3 

unlist(sapply(slot(ooo, "polygons"), function(p) sapply(slot(p, "Polygons"), slot, "area")))

#> [1] 1700211.935 7269743.131  311777.515    1212.778

set_RGEOS_polyThreshold(oT)
set_RGEOS_warnSlivers(oW)
set_RGEOS_dropSlivers(oD)
# see thread https://stat.ethz.ch/pipermail/r-sig-geo/2015-September/023468.html
Pol1=rbind(c(0,0),c(0,10),c(10,10),c(10,0),c(0,0))
Pol2=rbind(c(0,0),c(10,0),c(10,-10),c(0,-10),c(0,0))
library(sp)
Pols1=Polygons(list(Polygon(Pol1)),"Pols1")
Pols2=Polygons(list(Polygon(Pol2)),"Pols2")
MyLay=SpatialPolygons(list(Pols1,Pols2))
Pol1l=Pol1+0.5
Pol2l=Pol2+0.5
Pols1l=Polygons(list(Polygon(Pol1l)),"Pols1l")
Pols2l=Polygons(list(Polygon(Pol2l)),"Pols2l")
MyLayl=SpatialPolygons(list(Pols1l,Pols2l))
inter=gIntersection(MyLay, MyLayl)
plot(MyLay)

#> Error in as.double(y): cannot coerce type 'S4' to vector of type 'double'

plot(MyLayl,add=TRUE)

#> Error in as.double(y): cannot coerce type 'S4' to vector of type 'double'

plot(inter, add=TRUE, border="green")

#> Error in as.double(y): cannot coerce type 'S4' to vector of type 'double'

try(gIntersection(MyLay, MyLayl, unaryUnion_if_byid_false=FALSE))

#> Error in RGEOSBinTopoFunc(spgeom1, spgeom2, byid, id, drop_lower_td, unaryUnion_if_byid_false,  : 
#>   TopologyException: no outgoing dirEdge found at 0.5 0

Pol1=rbind(c(0,0),c(0,1),c(1,1),c(1,0),c(0,0))
Pol2=rbind(c(0,0),c(1,0),c(1,-1),c(0,-1),c(0,0))
Pols1=Polygons(list(Polygon(Pol1)),"Pols1")
Pols2=Polygons(list(Polygon(Pol2)),"Pols2")
MyLay=SpatialPolygons(list(Pols1,Pols2))
Pol1l=Pol1+0.1
Pol2l=Pol2+0.1
Pols1l=Polygons(list(Polygon(Pol1l)),"Pols1l")
Pols2l=Polygons(list(Polygon(Pol2l)),"Pols2l")
MyLayl=SpatialPolygons(list(Pols1l,Pols2l))
inter=gIntersection(MyLay, MyLayl, unaryUnion_if_byid_false=FALSE)
gEquals(inter, MyLay)

#> [1] TRUE

inter1=gIntersection(MyLay, MyLayl, unaryUnion_if_byid_false=TRUE)
gEquals(inter1, MyLay)

#> [1] FALSE

gEquals(inter, inter1)

#> [1] FALSE

plot(MyLay, ylim=c(-1, 1.1))

#> Error in as.double(y): cannot coerce type 'S4' to vector of type 'double'

plot(MyLayl, add=TRUE)

#> Error in as.double(y): cannot coerce type 'S4' to vector of type 'double'

plot(inter, angle=45, density=10, add=TRUE)

#> Error in as.double(y): cannot coerce type 'S4' to vector of type 'double'

plot(inter1, angle=135, density=10, add=TRUE)

#> Error in as.double(y): cannot coerce type 'S4' to vector of type 'double'

inter2=gIntersection(MyLay, MyLayl)
gEquals(inter2, MyLay)

#> [1] FALSE

gEquals(inter1, inter2)

#> [1] TRUE