gCentroid {rgeos}	R Documentation

Centroid of Geometry

Description

Function calculates the centroid of the given geometry.

Usage

gCentroid(spgeom, byid=FALSE, id = NULL)

Arguments

spgeom
sp object as defined in package sp

byid
Logical determining if the function should be applied across subgeometries (TRUE) or the entire object (FALSE)

id
Character vector defining id labels for the resulting geometries, if unspecified returned geometries will be labeled based on their parent geometries' labels.

Details

Returns a SpatialPoints object of the centroid(s) for spgeom.

Author(s)

Roger Bivand & Colin Rundel

See Also

gBoundary gConvexHull gEnvelope gPointOnSurface

Examples

x = readWKT(paste("GEOMETRYCOLLECTION(POLYGON((0 0,10 0,10 10,0 10,0 0)),",
 "POLYGON((15 0,25 15,35 0,15 0)))"))

# Centroids of both the square and circle independently
c1 = gCentroid(x,byid=TRUE) 
# Centroid of square and circle together
c2 = gCentroid(x)

plot(x)
plot(c1,col='red',add=TRUE)
plot(c2,col='blue',add=TRUE)

[Package rgeos version 0.5-2 Index]