bbox2SP {rgeos}	R Documentation

Converts a bounding box into a SpatialPolygons object.

Description

Converts a bounding box into a SpatialPolygons object.

Usage

bbox2SP(n,s,w,e,bbox=NA,proj4string=CRS("+init=epsg:4326"))

Arguments

n
the top north latitude

s
the bottom south latitude

w
the most western longitude

e
the most eastern longitude

bbox
a bounding box 2 x 2 matrix as produced by bbox

proj4string
a coordinate reference system as defined in CRS

Details

This function converts a set of coordinates limiting a bounding box into a SpatialPolygons. It can be used for instance to clip a subset of a larger spatial object (e.g. using gIntersection)

Value

An object of SpatialPolygons class.

Examples

if (require(rgdal)) {
cities <- readOGR(dsn=system.file("vectors", package = "rgdal")[1], layer="cities")
n<-75
s<-30
w<--40
e<-32
myPoly<-bbox2SP(n,s,e,w)
plot(cities)
plot(myPoly,border="red",add=TRUE)

bb<-bbox(cities)
myPoly<-bbox2SP(bbox=bb,proj4string=CRS(proj4string(cities)))
plot(myPoly,add=TRUE,border="blue")
}

[Package rgeos version 0.4-1 Index]