create SpatialRings or SpatialRingsDataFrame

create objects of class SpatialRings or SpatialRingsDataFrame

Ring(coords,ID=as.character(NA))
SpatialRings(RingList, proj4string=CRS(as.character(NA)))
SpatialRingsDataFrame(sr, data, match.ID = TRUE)

Arguments

coords
2-column numeric matrix with coordinates; first point (row) should equal last coordinates (row); if the hole argument is not given, the status of the polygon as a hole or an island will be taken from the ring direction, with clockwise meaning island, and counter-clockwise meaning hole

ID
character vector of length one with identifier

RingList
list with objects of class Ring-class

proj4string
Object of class "CRS" holding a valid proj4 string

sr
object of class SpatialRings-class

data
object of class data.frame; the number of rows in data should equal the number of Lines elements in sl

match.ID
logical: (default TRUE): match SpatialLines member Lines ID slot values with data frame row names, and re-order the data frame rows if necessary

Value

Ring returns object of class Ring SpatialRings returns object of class SpatialRings SpatialRingsDataFrame returns object of class SpatialRingsDataFrame

See also